pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.14.0"
requests = "^2.26.0"
gunicorn = "^20.1.0"
watchdog = "^2.1.7"
Flask = "^2.1.0"
oauthlib = "^3.2.0"
Flask-Login = "^0.6.1"
pymongo = "^4.1.1"
mongomock = "^4.1.2"
furl = "^2.1.3"
Flask-User = "^1.0.2"
Flask-Principal = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
